smart-bridge (varadhi)
In this repository, we have integrated all the necessary codes and a working video of our project. Our primary objective is to prioritize the safety of human lives and ensure secure transportation during flood events in developing regions.

We are committed to making flood-prone areas safer for everyone. Your contributions and feedback are greatly appreciated 🌟