abstract
**PROBLEM STATEMENT:**
Floods lead to a vast loss of life and property in many countries. But in developing countries the lack of proper technology leads to more loss of life in many situations. Bridge failures are one of the most infrastructure problems in the world. In rainy season due to heavy rains the level of flood increases and it's also blocks the way and people cannot travel to the other side of the bridge. It often leads to the severe consequences, loss of life, restricted commerce. Whenever there is a disaster there is loss of lives, damage to the public property. The main objective of this project is to monitor the flood situation and  lift the bridge in case of danger. 

**THE APPROACH TO SOLVE THE PROBLEM:**
 In order to overcome this problem, we have to measure flood-level of the water in the river by using a moisture sensor. That output is connected to an Arduino microcontroller. Due to heavy rains whenever the water level increases and touches the danger-level the controller gets data from the sensor and alerts the people by buzzer and at the ends of the bridge gates are closed automatically with the help of servo motor .Then it lifts the bridge by the help of hydraulic cylinder until the water level goes below the danger level and after the bridge is lifted the gates are opened so that the vehicles can reach the other side of the bridge without any loss.